Fix size types in function and block maps

+++ include/viua/loader.h
@@ -22,10 +22,10 @@ class Loader {
 
     std::vector<unsigned> jumps;
 
-    std::map<std::string, uint16_t> function_addresses;
+    std::map<std::string, uint64_t> function_addresses;
     std::map<std::string, unsigned> function_sizes;
     std::vector<std::string> functions;
-    std::map<std::string, uint16_t> block_addresses;
+    std::map<std::string, uint64_t> block_addresses;
     std::vector<std::string> blocks;
 
     IdToAddressMapping loadmap(char*, const uint16_t&);
@@ -45,11 +45,11 @@ class Loader {
 
     std::vector<unsigned> getJumps();
 
-    std::map<std::string, uint16_t> getFunctionAddresses();
+    std::map<std::string, uint64_t> getFunctionAddresses();
     std::map<std::string, unsigned> getFunctionSizes();
     std::vector<std::string> getFunctions();
 
-    std::map<std::string, uint16_t> getBlockAddresses();
+    std::map<std::string, uint64_t> getBlockAddresses();
     std::vector<std::string> getBlocks();
 
     Loader(std::string pth): path(pth), size(0), bytecode(nullptr) {}
